+++ rs/rosetta-api/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- /account/balance endpoint can now return information about neurons.
+
 ## [1.2.0] - 2021-09-28
 ### Changed
 - Neuron address derivation is now supported by /construction/derive endpoint.